+++ docs/sphinx/conf.py
@@ -127,7 +127,8 @@ intersphinx_mapping = {'python': ('https://docs.python.org/3.7', None),
                        'astropy': ('http://docs.astropy.org/en/latest', None),
                        'numpy': ('http://docs.scipy.org/doc/numpy/', None),
                        'sdss_access': ('https://sdss-tree.rtfd.io/en/latest', None),
-                       'sdss_tree': ('https://sdss-access.rtfd.io/en/latest', None)}
+                       'sdss_tree': ('https://sdss-access.rtfd.io/en/latest', None),
+                       'sdssdb': ('https://sdssdb.rtfd.io/en/latest', None)}
 
 autodoc_mock_imports = ['_tkinter']
 autodoc_member_order = 'groupwise'