+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Change Log
 
 * :release:`0.1.1 <unknown>`
 * :feature:`-` added decorators, `access_loader`, `parser_loader, `sdss_loader` to reduce boilerplate for overriding abstract methods
+* :feature:`-` added new ``BrainNoAccess`` class to support non sdss_access paths
 * :feature:`-` added new ``MMAccess`` mixin combining ``AccessMixIn`` and ``MMAMixIn``
 * :feature:`-` split out `sdss_access` parts into new ``AccessMixIn``.
 * :bug:`2` fixed conflict between loading files and `sdss_access` necessary path parameters